Matrix now uses long int to avoid overflow with really big data matrices
git-svn-id: http://cuda-convnet.googlecode.com/svn/trunk@361 bc73d74b-de6f-cad7-86a4-6ddd3d5d919c

+++ include/data.cuh
@@ -78,7 +78,7 @@ protected:
     CPUData* _hData;
     NVMatrixV _data;
     int _minibatchSize;
-    int _dataSize;
+    long int _dataSize;
 public:
     DataProvider(int minibatchSize);
     GPUData& operator[](int idx);
